wristband_control/gesture_recognition/requirements.txt
numpy>=1.21.0
scikit-learn>=1.0.0
joblib>=1.1.0
pyserial>=3.5
